+++ app/components/LoadingIndicator.js
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+
+class LoadingIndicator extends Component {
+
+	render() {
+		return (
+			<p>Fetching...</p>
+		)
+	}
+}
+
+export default LoadingIndicator;