+++ app/config/production.js
@@ -1,0 +1,6 @@
+import assign from 'object-assign';
+import sharedConfig from './shared';
+
+export default assign({
+	
+}, sharedConfig);