+++ app/containers/App.js
@@ -1,7 +1,7 @@
-import React, { Component } from 'react'
-import { Link } from 'react-router'
-import Helmet from 'react-helmet'
-import './scss/App.scss'
+import React, { Component } from 'react';
+import { Link } from 'react-router';
+import Helmet from 'react-helmet';
+import './scss/App.scss';
 
 class App extends Component {
 
@@ -18,8 +18,7 @@ class App extends Component {
 				/>
 				<nav>
 					<ul>
-						<li><Link to='/'>Home</Link></li>
-						<li><Link to='/users'>Users</Link></li>
+						<li><Link to='/'>Users</Link></li>
 					</ul>
 				</nav>
 				<div className='route-content'>