+++ app/index.js
@@ -1,1 +1,7 @@
-import './Router'
+import 'isomorphic-fetch';
+
+if (CONFIG.mockApi) {
+	require('./mocks');
+}
+
+export serverMiddleware from './Router';