+++ app/mocks/index.js
@@ -1,0 +1,2 @@
+import './users';
+import './user';