+++ app/reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import users from './users';
+import user from './user';
+
+export default combineReducers({
+	users,
+	user
+});