+++ server.js
@@ -1,31 +1,35 @@
-require('babel-core/register')({
-	stage: 0
-});
+require('babel-core/register');
 
 require.extensions['.scss'] = function() {
 	return;
 };
 
-var path = require('path');
-var express = require('express');
-var webpack = require('webpack');
-var webpackMiddleware = require('webpack-dev-middleware');
-var webpackHotMiddleware = require('webpack-hot-middleware');
-var config = require('./webpack.config.js');
-var Router = require('./app/Router');
+const fs = require('fs');
+const path = require('path');
+const express = require('express');
+const webpack = require('webpack');
+const webpackMiddleware = require('webpack-dev-middleware');
+const webpackHotMiddleware = require('webpack-hot-middleware');
+const config = require('./webpack.config.js');
 
 const isDeveloping = process.env.NODE_ENV !== 'production';
 const port = isDeveloping ? 3000 : process.env.PORT;
-const app = express();
+const server = express();
 
-app.use(express.static(path.resolve(__dirname, 'dist')));
+// Short-circuit the browser's annoying favicon request. You can still
+// specify one as long as it doesn't have this exact name and path.
+server.get('/favicon.ico', function(req, res) {
+	res.writeHead(200, { 'Content-Type': 'image/x-icon' });
+	res.end();
+});
+
+server.use(express.static(path.resolve(__dirname, 'dist')));
 
 if (isDeveloping) {
 	const compiler = webpack(config);
 
-	app.use(webpackMiddleware(compiler, {
+	server.use(webpackMiddleware(compiler, {
 		publicpath: config.output.publicpath,
-		contentBase: 'app',
 		watch: true,
 		stats: {
 			colors: true,
@@ -37,14 +41,18 @@ if (isDeveloping) {
 		}
 	}));
 
-	app.use(webpackHotMiddleware(compiler, {
+	server.use(webpackHotMiddleware(compiler, {
 		path: '/__webpack_hmr'
 	}));
 }
 
-app.get('*', Router);
+// The client has access to the same global `CONFIG` variable.
+const envConfigPath = './app/config/' + (process.env.NODE_ENV || 'local');
+global.CONFIG = require(envConfigPath).default;
 
-app.listen(port, 'localhost', function onStart(err) {
+server.get('*', require('./app').serverMiddleware);
+
+server.listen(port, 'localhost', function onStart(err) {
 	if (err) {
 		console.log(err);
 	}