+++ CHANGELOG.md
@@ -15,3 +15,8 @@
 
 0.2.3
 	- Make bundle path absolute so it works for deep urls
+
+1.0.0-alpha
+	- Use Redux, the simplicity was nice but to scale a central store to dispatch to makes good sense.
+	- Shared app config.
+	- Data mocking.